adding mailgun handler and a beginning on setting up the failover logic for multiple handlers

+++ workers/worker.py
@@ -2,7 +2,19 @@ import logging
 import boto3
 import os
 import json
-from mail_handlers import ses_handler
+from mail_handlers import ses_handler, mailgun_handler
+
+
+def setup_handlers():
+    '''
+    Factory method for setting up handlers for the worker
+    '''
+    handlers = []
+    ses = ses_handler.SESHandler()
+    handlers.append(ses)
+    mailgun = mailgun_handler.MailgunHandler()
+    handlers.append(mailgun)
+    return handlers
 
 
 def process_messages():
@@ -10,14 +22,16 @@ def process_messages():
     sqs = boto3.resource('sqs', region_name=os.environ.get('REGION'))
     queue = sqs.get_queue_by_name(QueueName=os.environ.get('QUEUE'))
 
-    ses = ses_handler.SESHandler()
+    handlers = setup_handlers()
+
     while True:
-        logging.warning('Waiting for messages')
+        logging.info('Waiting for messages')
         messages = queue.receive_messages(WaitTimeSeconds=20, MaxNumberOfMessages=10)
         for message in messages:
             messsage_body = json.loads(message.body)
-            success = ses.send_mail(messsage_body)
-            message.delete()
+            
+            if success:
+                message.delete()
 
 
 if __name__ == '__main__':
